Move get_config and get_secret into a z2jh module
This serves multiple purposes:

- Allows subconfigs loaded via hub.extraConfig to use get_config
  and get_secret
- Stops overloading the `get_config` method that is injected in
  config files by traitlets

Fixes #374
